+++ lib/ricordami/can_be_serialized.rb
@@ -1,0 +1,12 @@
+require "active_model"
+
+module Ricordami
+  module CanBeSerialized
+    extend ActiveSupport::Concern
+
+    included do
+      include ActiveModel::Serializers::JSON
+      include ActiveModel::Serializers::Xml
+    end
+  end
+end